Added cssnext with vars and imports
This has gotten very close to an actual sass setup. It includes using
css custom properties to define variables in one file and then importing
that file into other files to use the vars.

Notes:
- The order of the loaders can cause things to not work. For instance,
having the postcss loader run before the cssnext loader was causing
the cssnext loader to not work with imports or vars. Playing around with
the order is not always intuitive.
- Some loaders don't seem to work well together. Originally I was only
using postcss and adding plugins on top of local scope, but this was
just not working. So I added cssnext loader and it worked.
- A little confusing to know that I had to put from in the cssnext
  options to get imports working.
- I like how you import from a root rather then having to ../.

TODO
- does importing the vars file multiple times cause duplicated (probably
yes but there could also be a task to remove duplicate css)