React es6 full examples with local styles
Successful setup of react with es6 and the local stles.

Notes:
- Getting babel working was difficult likely because of old
documentation. It was hard to get it to find the correct files, it was
previously not finding the JS files to compile and complaining. I think
the issue was using the babel-loader rather then just `babel`. This
could mean problems for webpack in general.
- Getting style imports working with es6 syntax was all really easy.
- Adding class syntax for es6 was really easy.
- There was something about hot loading react to change it dynamically
rather then server reloading. This might help.

TODO
- Full accordion example.